^R Finish hostutil split: move remaining helpers into internal/host/launcher and delete internal/hostutil
Sixth and final sub-PR of the hostutil split. Moves the path
canonicalization (CanonicalizePath, CanonicalizePathFrom, RealHome,
CanonicalizeToolPath) and the launcher runtime helpers
(ExtractCodexVersion, ValidateSecurityOptions, DedupeEndpointList,
ResolveEndpoints) from internal/hostutil/ into internal/host/launcher/,
then deletes internal/hostutil/ entirely.

Git detects these as renames so the diff is minimal: hostutil.go ->
launcher/paths.go (with expandUser collapsed into a direct
pathutil.ExpandUserPathBestEffort call), hostutil_test.go ->
launcher/paths_test.go (package rename only), launcher.go ->
launcher/runtime.go (kept verbatim modulo package name).

cmd/workcell-hostutil/main.go drops its hostutil import; the only
remaining `hostutil` tokens in the repo refer to the binary name
(workcell-hostutil, go_hostutil bash function) which is unrelated.

Also refreshes a few doc comments and design-doc paths that named the
old internal/hostutil package or its files. go build/test/gofmt clean;
no behavior change.
