--- internal/hostutil/doc.go
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright 2026 Omkhar Arasaratnam
-
-// Package hostutil provides host-side helpers for Workcell tooling.
-package hostutil
